requirements.txt
flask>=3.0.0
flask-login>=0.6.3
gunicorn>=21.2.0
numpy>=1.26.3
opencv-python>=4.9.0.80
torch>=2.2.0
torchvision>=0.17.0
prometheus-client>=0.19.0
python-json-logger>=2.0.7
pyyaml>=6.0.1
google-cloud-storage>=2.14.0
pytest>=8.0.0
pytest-cov>=4.1.0
pylint>=3.0.3
black>=24.1.1
isort>=5.13.2
